+++ Kernel/cpan-lib/Sisimai/Data.pm
@@ -63,7 +63,7 @@ sub new {
     };
 
     # Create Sisimai::Time object
-    $thing->{'timestamp'} = gmtime Sisimai::Time->new($argvs->{'timestamp'});
+    $thing->{'timestamp'} = Sisimai::Time->new($argvs->{'timestamp'});
     $thing->{'timezoneoffset'} = $argvs->{'timezoneoffset'} // '+0000';
 
     # Callback method
@@ -187,7 +187,7 @@ sub make {
                 $p->{'timestamp'} = ($t->epoch - $zoneoffset) // undef;
             };
         }
-        next unless $p->{'timestamp'};
+        next unless defined $p->{'timestamp'};
 
         OTHER_TEXT_HEADERS: {
             # Scan "Received:" header of the original message
@@ -251,13 +251,18 @@ sub make {
                     if( $vm > 2 ) {
                         # Build regular expression for removing string like '550-5.1.1'
                         # from the value of "diagnosticcode"
-                        my $re = qr/[ ]$vr[- ](?:\Q$vs\E)?/;
+                        my $re0 = qr/;?[ ]$vr[- ](?:\Q$vs\E)?/;
+                        my $re1 = qr/;?[ ][45]\d\d[- ](?:\Q$vs\E)?/;
+                        my $re2 = qr/;?[ ]$vr[- ](?:[45][.]\d[.]\d+)?/;
 
                         # 550-5.7.1 [192.0.2.222] Our system has detected that this message is
                         # 550-5.7.1 likely unsolicited mail. To reduce the amount of spam sent to Gmail,
                         # 550-5.7.1 this message has been blocked. Please visit
                         # 550 5.7.1 https://support.google.com/mail/answer/188131 for more information.
-                        $p->{'diagnosticcode'} =~ s/$re/ /g;
+                        $p->{'diagnosticcode'} =~ s/$re0/ /g;
+                        $p->{'diagnosticcode'} =~ s/$re1/ /g;
+                        $p->{'diagnosticcode'} =~ s/$re2/ /g;
+                        $p->{'diagnosticcode'} =~ s|<html>.+</html>||i;
                         $p->{'diagnosticcode'} =  Sisimai::String->sweep($p->{'diagnosticcode'});
                     }
                 }
@@ -722,7 +727,7 @@ azumakuniyuki
 
 =head1 COPYRIGHT
 
-Copyright (C) 2014-2020 azumakuniyuki, All rights reserved.
+Copyright (C) 2014-2020,2022 azumakuniyuki, All rights reserved.
 
 =head1 LICENSE