+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 6.5.3 2023-??-??
  - 2023-05-31 Process management: Added missing scope filter for transition actions.
+ - 2023-05-30 Updated CPAN library Sisimai to version 4.25.16. [#391](https://github.com/znuny/Znuny/issues/391)
  - 2023-05-25 Changed article action 'note to linked ticket' to only be visible/usable in agent ticket zoom.
  - 2023-05-24 Selected value in a dynamic field web service dropdown field can now be removed with 'x' button.
  - 2023-05-23 Fixed event handling for DBCRUD history entries.